WIP adding scaffolding for EmbargoTerminationApproval
[skip ci]

[#OSF-5641]
