+++ website/mails/mails.py
@@ -199,11 +199,19 @@ PENDING_REGISTRATION_NON_ADMIN = Mail(
     'pending_registration_non_admin',
     subject='Registration pending for one of your projects.'
 )
+PENDING_EMBARGO_TERMINATION_ADMIN = Mail(
+    'pending_embargo_termination_admin',
+    subject='Registration pending for one of your projects.'
+)
+PENDING_EMBARGO_TERMINATION_NON_ADMIN = Mail(
+    'pending_embargo_termination_non_admin',
+    subject='Registration pending for one of your projects.'
+)
+
 FILE_OPERATION_SUCCESS = Mail(
     'file_operation_success',
     subject='Your ${action} has finished',
 )
-
 FILE_OPERATION_FAILED = Mail(
     'file_operation_failed',
     subject='Your ${action} has failed',
